SIUUUUUU

+++ db/db_operations.py
@@ -10,7 +10,7 @@ password = credential_dict.get("password")
 host = credential_dict.get("host")
 port = credential_dict.get("port")
 
-def save_score(client: MongoClient, score: int) -> None:
+def save_score(client: MongoClient, score: int, name: str) -> None:
     """
     Saves the score to the 'highscores' collection in MongoDB.
 
@@ -25,11 +25,11 @@ def save_score(client: MongoClient, score: int) -> None:
         highscores_col = check_and_create_collection(client)
         num_documents = highscores_col.count_documents({})
         lowest_score_doc = highscores_col.find().sort('score', pymongo.ASCENDING).limit(1)
-        if num_documents < 5:
-            data = {"user": "test_mule", "score": score}
+        if num_documents < 5 and name != "":
+            data = {"user": name, "score": score}
             highscores_col.insert_one(data)
-        if score > lowest_score_doc[0]['score'] and num_documents >= 5:
-            data = {"user": "test_mule", "score": score}
+        if score > lowest_score_doc[0]['score'] and num_documents >= 5 and name != "":
+            data = {"user": name, "score": score}
             highscores_col.insert_one(data)
             highscores_col.delete_one({"_id": lowest_score_doc[0]['_id']}) #Deleting the worst entry
     except Exception as e:
@@ -58,6 +58,18 @@ def get_high_scores(client: MongoClient) -> list:
         return sorted_scores
     except Exception as e:
         print(f"Error retrieving high scores: {e}")
+
+def get_worst_score_in_db(client: MongoClient):
+        db = client["FlappyMule"]
+        highscores_col = db['FlappyMuleScores']
+        sorted_scores = highscores_col.find().sort('score', pymongo.DESCENDING).limit(5)
+        worst_val = 0
+        num_of_docs = 0
+        for score_doc in sorted_scores:
+            score_value = score_doc.get('score')
+            worst_val = score_value
+            num_of_docs += 1
+        return worst_val, num_of_docs
 
 def check_and_create_collection(client: MongoClient):
         mydb = client["FlappyMule"]
